Removed TypeScript
