+++ frontend/beerme/src/Components/Layout/Header.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import * as Constants from "../../Utils/Constants";
-import MainController from "../../Controllers/MainController.ts";
+import MainController from "../../Controllers/MainController";
 
 export function Header() {
   const onMouseOver = event => {
