+++ frontend/beerme/src/Utils/NetClient.js
@@ -1,11 +1,11 @@
 /**
  * Example GET request:
- * 
+ *
  *    NetClient.get("http://jsonplaceholder.typicode.com/todos")
  *      .then(res => res.json())
  *      .then(data => console.log(data))
  *      .catch(console.log);
- *      
+ *
  * Example POST request:
  *    NetClient.post("https://httpbin.org/post", { test: "hola" })
  *      .then(res => res.json())
@@ -21,8 +21,11 @@ class NetClient {
   static async post(url: string, data: Object) {
     return await fetch(url, {
       method: "POST",
-      body: JSON.stringify(data)
-    });
+      body: JSON.stringify(data),
+      headers: {
+        "Content-Type": "application/json"
+      }
+    }).catch(console.log);
   }
 }
